Refactor `behaves_like` implementation to fix spec
A reply is given for the [RSpec
issue](https://github.com/rspec/rspec-expectations/issues/1117), I've
reviewed the proposed solution and found that I'd take a different
approach instead in order to fix the `at_exit` spec.

First off, this commit extracts out the core logic for `behaves_like`
into a `private` method. This is important for a couple of reason.

By extracting it out, `behaves_like` now only consists of the `at_exit`
logic, and because I can't test `at_exit` due to the nature of it, I've
decided to test directly on the newly extracted `private` method.

While it is generally frowned upon to test private methods (you should
be testing the public interfaces only!!), and I do agree on that on a
general case, I think that this is an exceptional situation that calls
for a break of rules.

Guidelines should be all that is, guides, it should not be rules that
prevent innovation. Thus I've decided to break the rule this time.

Fixes #2